+++ CHANGELOG.md
@@ -19,7 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Make `B2Api.get_bucket_by_id` return populated bucket objects in v2
 * Add proper support of `recommended_part_size` and `absolute_minimum_part_size` in `AccountInfo`
 * Refactored `minimum_part_size` to `recommended_part_size` (tha value used stays the same)
+* Refactored `sync.file.*File` and `sync.file.*FileVersion` to `sync.path.*SyncPath`
 * Encryption settings, types and providers are now part of the public API
+* Refactored `FileVersionInfo` to `FileVersion`
+* `Bucket` unit tests for v1 and v2 are now common
 
 ### Removed
 * Remove `Bucket.copy_file` and `Bucket.start_large_file`